backend/requirements.txt
Django==4.1.7
django-debug-toolbar==3.8.1
django-filter==22.1
djangorestframework==3.14.0
elasticsearch-dsl==7.4.0
psycopg2-binary==2.9.5